backend/requirements-main.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
alembic==1.16.1
amqp==5.3.1
annotated-types==0.7.0
anthropic==0.75.0
anyio==4.9.0
argcomplete==1.10.3
asgiref==3.8.1
astroid==3.3.10
async-timeout==4.0.3
asyncpg==0.30.0
snowflake-sqlalchemy>=1.4.0
snowflake-connector-python>=3.0.0
attrs==25.3.0
backoff==2.2.1
bcrypt==4.0.1
beautifulsoup4==4.8.2
billiard==4.2.1
boto3==1.42.5
botocore==1.42.5
build==1.2.2.post1
cachetools==5.5.2
celery==5.3.6
certifi==2025.4.26
cffi==2.0.0
chardet==3.0.4
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coloredlogs==15.0.1
compressed-rtf==1.0.7
coverage==7.7.1
croniter==6.0.0
cryptography==44.0.3
dataclasses-json==0.6.7
decorator==5.2.1
Deprecated==1.2.18
dill==0.4.0
distro==1.9.0
dnspython==2.7.0
docx2txt==0.9
dotenv==0.9.9
durationpy==0.10
ebcdic==1.1.1
ecdsa==0.19.1
email_validator==2.2.0
exceptiongroup==1.3.0
extract-msg==0.28.7
fastapi==0.115.11
fastapi-cache2==0.2.2
filelock==3.17.0
flake8==7.1.2
flatbuffers==25.2.10
flower==2.0.1
frozenlist==1.6.0
fsspec==2024.12.0
google-auth[requests]>=2.46.0,<3.0.0
googleapis-common-protos==1.70.0
greenlet==3.1.1
grpcio==1.71.2
h11==0.14.0
hf-xet==1.1.2
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.33.4
humanfriendly==10.0
humanize==4.12.3
idna==3.10
IMAPClient==2.1.0
importlib_metadata==8.6.1
importlib_resources==6.5.2
iniconfig==2.1.0
isort==6.0.1
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
json-repair==0.44.1
jsonpatch==1.33
jsonpointer==3.0.0
junit-xml-2==1.9
kombu==5.5.1
kubernetes==32.0.1
langchain==1.2.3
langchain-anthropic==1.3.1
langchain-chroma==1.1.0
langchain-core==1.2.7
langchain-huggingface==1.2.0
langchain-ollama==1.0.1
langchain-openai==1.1.7
langchain-text-splitters==1.1.0
langchain-aws==1.2.0
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.1
langsmith==0.5.0
loguru==0.7.3
lxml==5.4.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mccabe==0.7.0
mdurl==0.1.2
mmh3==5.1.0
mpmath==1.3.0
multidict==6.4.4
mypy_extensions==1.1.0
networkx==3.4.2
numpy==2.2.6
oauthlib==3.2.2
olefile==0.47
ollama==0.6.1
onnxruntime==1.20.0
openai>=1.65.5
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-instrumentation==0.57b0
opentelemetry-instrumentation-asgi==0.57b0
opentelemetry-instrumentation-fastapi==0.57b0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
opentelemetry-util-http==0.57b0
orjson==3.10.18
ormsgpack==1.12.1
overrides==7.7.0
packaging==24.2
passlib==1.7.4
pdfminer.six==20191110
pendulum==3.1.0
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.6.0
posthog==4.2.0
prometheus_client==0.22.0
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==5.29.5
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.12.1
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.9
pydantic-settings==2.11.0
pydantic_core==2.33.2
pyflakes==3.2.0
Pygments==2.19.1
pylint==3.3.6
pylint-fail-under==0.3.0
pylint_junit==0.3.5
pypdf==5.6.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-cov==6.0.0
pytest-env==1.1.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose[cryptography]==3.5.0
python-multipart==0.0.20
python-pptx==0.6.23
pytz==2025.1
PyYAML==6.0.2
redis==6.2.0
regex==2024.11.6
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rsa==4.9.1
s3transfer==0.16.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
sentence-transformers==4.1.0
shellingham==1.5.4
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SpeechRecognition==3.8.1
SQLAlchemy==2.0.38
starlette==0.46.1
starlette-context==0.4.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
tomli==2.2.1
tomlkit==0.13.2
torch==2.9.1
torchaudio==2.9.1
tornado==6.5.1
tqdm==4.67.1
transformers==4.52.4
typer==0.16.0
types-PyYAML==6.0.12.20250516
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.1
tzlocal==5.3.1
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
vine==5.1.0
watchfiles==1.0.4
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
wrapt==1.17.2
xlrd==1.2.0
XlsxWriter==3.2.3
xxhash==3.5.0
yarl==1.20.0
zipp==3.22.0
zstandard==0.23.0
pdf2image==1.17.0
smbprotocol==1.15.0
azure-storage-blob==12.27.1
google-cloud-speech==2.34.0
google-cloud-storage==3.5.0
google-cloud-recaptcha-enterprise==1.26.0
mcp>=1.0.0